backend/requirements.txt
asgiref==3.8.1
chowkidar-strawberry==0.3.2
Django==4.2.19
django-ipware==7.0.1
djangorestframework==3.15.2
graphql-core==3.2.6
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-ipware==3.0.0
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.260.2
typing_extensions==4.12.2
